impl(005): T007-T014 — snapshot MVP: binary COPY -> Arrow, conformance green
copy_decode.rs: incremental binary-COPY parser straight into Arrow
builders (numeric NBASE->i128 at declared scale with drift checks,
epoch rebase with infinity saturation, uuid/jsonb text forms, batch
cuts on row boundaries only; total — typed errors, no panics).
sqlgen.rs: strict ident quoting + server-side policy projections.
Source impl: once-per-run cached reflection, structured pushes under
byte-budget backpressure, S4 cancellation, empty tables materialize
via a zero-row schema batch. CLI/facade wired (postgres-source
feature). Nets: 36 tests green — conformance matrix (every contract
row asserted through real pg -> DuckDB, incl. hostile identifiers,
±infinity ordering, dest ident normalization), differential proptest
(decoder byte-identical to an independent FromSql reference), fuzz
target pg_copy_decode registered. Quickstart smoke: 1000 rows via
release CLI in 18 ms.

+++ crates/rdlt-source-postgres/src/types.rs
@@ -283,7 +283,7 @@ mod tests {
         let typmod = ((50 << 16) | 2) + 4;
         assert_eq!(map_type(&base(oid::NUMERIC, typmod)).select, SelectPolicy::CastText);
         // Negative scale (PG15+) → text.
-        let typmod = ((10 << 16) | (((-2i32) & 0x7FF) as i32)) + 4;
+        let typmod = ((10 << 16) | ((-2i32) & 0x7FF)) + 4;
         assert_eq!(map_type(&base(oid::NUMERIC, typmod)).select, SelectPolicy::CastText);
         // Enum → label text.
         let e = map_type(&PgTypeInfo { oid: 99_999, typtype: 'e', typcategory: 'E', typmod: -1 });
